Order-service
Atualização Order, Products, User-service

+++ user-service/src/infrastructure/database/repositories/user-prisma.repository/user-prisma.repository.service.ts
@@ -31,52 +31,101 @@ export class UserPrismaRepository implements IUserRepository {
   constructor(private readonly prisma: PrismaService) {}
 
   async findById(id: number) {
-    return await this.prisma.user.findUnique({
-      where: { id },
-      select: this.userSelect,
-    });
+
+    try{
+        const user = await this.prisma.user.findUnique({
+          where: { id },
+          select: this.userSelect,
+        });
+
+      if (!user) {
+        throw new NotFoundException(`Usuário com ID ${id} não encontrado.`);
+      }
+      return user;
+
+    }
+    catch (error) {
+      if (error instanceof Prisma.PrismaClientKnownRequestError && error.code === 'P2025') {
+        throw new NotFoundException(`Usuário com ID ${id} não encontrado.`);
+      }
+    }
   }
 
   async findByEmail(email: string) {
-     return await this.prisma.user.findUnique({
-      where: { email },
-      select: this.userSelect,
-    });
+
+    try{
+        const findemail = await this.prisma.user.findUnique({
+          where: { email },
+          select: this.userSelect,
+        });
+
+        if (!findemail) {
+          throw new NotFoundException(`Usuário com email ${email} não encontrado.`);  
+        }
+        return findemail;
+    }
+    catch (error) {
+        if (error instanceof Prisma.PrismaClientKnownRequestError && error.code === 'P  2025') {
+          throw new NotFoundException(`Usuário com email ${email} não encontrado.`);  
+        }
+    }
+     
   }
 
   async findAll() {
-    return await this.prisma.user.findMany({
-      select: this.userSelect,
-    });
+    try{
+         return await this.prisma.user.findMany({
+          select: this.userSelect,
+        });
+    }
+    catch (error) {
+      if (error instanceof Prisma.PrismaClientKnownRequestError && error.code === 'P2025') {
+        throw new NotFoundException(`Nenhum usuário encontrado.`);
+      }
+    } 
   }
 
   async create(data: CreateUserDto) {
     
-    const { address, ...userData } = data;
+    try{
+      const { address, ...userData } = data;
 
-    return await this.prisma.user.create({
-      data: {
-        ...userData,
-        addresses: {
-          create: [
-            {
-              street: address.street,
-              number: address.number,
-              cep: address.cep,
-              complement: address.complement,
-              city: {
-                connect: { id: address.cityId },
+      return await this.prisma.user.create({
+        data: {
+          ...userData,
+          addresses: {
+            create: [
+              {
+                street: address.street,
+                number: address.number,
+                cep: address.cep,
+                complement: address.complement,
+                city: {
+                  connect: { id: address.cityId },
+                },
               },
-            },
-          ],
+            ],
+          },
         },
-      },
-      select: this.userSelect,
-    });
+        select: this.userSelect,
+      });
+    }
+    catch (error) {
+      if (error instanceof Prisma.PrismaClientKnownRequestError && error.code === 'P2002') {
+        throw new NotFoundException(`Usuário com email ${data.email} já existe.`);
+      }
+      throw error;
+    } 
   }
 
   async update(id: number, data: UpdateUserDto) {
     try {
+      const finduser = await this.prisma.user.findUnique({
+        where: { id },
+      });
+      if (!finduser) {
+        throw new NotFoundException(`Usuário com ID ${id} não encontrado.`);
+      } 
       return await this.prisma.user.update({
         where: { id },
         data: data,
@@ -92,6 +141,12 @@ export class UserPrismaRepository implements IUserRepository {
 
   async delete(id: number): Promise<void> {
     try {
+      const user = await this.prisma.user.findUnique({
+        where: { id },
+      });
+      if (!user) {
+        throw new NotFoundException(`Usuário com ID ${id} não encontrado.`);
+      }
       await this.prisma.user.delete({
         where: { id },
       });
